Refactor code accessing and setting app paths
- Refactored 'subsearch/data/set_data.py' file into 'subsearch/data/data_initializer.py'.
- Renamed 'DownloadMetaData' -> 'DownloadData', 'FormattedMetadata' -> 'PrettifiedDownloadData'.
- Moved app paths into 'AppPaths' class as 'tmpdir' and 'appdata_local' attributes.
- Extracted SUPPORTED_FILE_EXTENSIONS and SUPPORTED_PROVIDERS from app configuration.
- Updated references to new app paths within 'subsearch/data/data_initializer.py' file.
- Deleted `application_config.json` from `subsearch/data`, its now created in `%username%/AppData/Local/Subsearch` & is persistent after a update
- Other minor naming changes and documentation improvements.
