+++ src/subsearch/data/__init__.py
@@ -1,6 +1,6 @@
-from .data_objects import GUI_DATA
+from .data_objects import GUI_DATA, SUPPORTED_FILE_EXTENSIONS, SUPPORTED_PROVIDERS
 from .guid import __guid__
-from .set_data import app_paths, video_data
+from .data_initializer import app_paths, video_data
 from .version import __version__
 
-__all__ = [GUI_DATA, app_paths, video_data, __version__, __guid__]
+__all__ = [GUI_DATA, SUPPORTED_FILE_EXTENSIONS, SUPPORTED_PROVIDERS, app_paths, video_data, __version__, __guid__]
