--- src/subsearch/data/application_config.json
@@ -1,39 +1,0 @@
-{
-    "current_language": "english",
-    "subtitle_type": {
-        "hearing_impaired": true,
-        "non_hearing_impaired": true
-    },
-    "percentage_threshold": 90,
-    "rename_best_match": true,
-    "context_menu": true,
-    "context_menu_icon": true,
-    "manual_download_fail": true,
-    "manual_download_mode": false,
-    "use_threading": true,
-    "multiple_app_instances": false,
-    "show_terminal": false,
-    "log_to_file": false,
-    "file_extensions": {
-        ".avi": true,
-        ".mp4": true,
-        ".mkv": true,
-        ".mpg": true,
-        ".mpeg": true,
-        ".mov": true,
-        ".rm": true,
-        ".vob": true,
-        ".wmv": true,
-        ".flv": true,
-        ".3gp": true,
-        ".3g2": true,
-        ".swf": true,
-        ".mswmm": true
-    },
-    "providers": {
-        "subscene_site": true,
-        "opensubtitles_site": true,
-        "opensubtitles_hash": true,
-        "yifysubtitles_site": true
-    }
-}
